Update the floating precision to support double.

+++ Compiler/include/Symbols.h
@@ -8,10 +8,10 @@ class Symbols {
 public:
 	Symbols();
 
-	typedef std::unordered_map<std::string, std::variant<float, int, bool>> SymbolsMap;
+	typedef std::unordered_map<std::string, std::variant<double, int, bool>> SymbolsMap;
 
 	SymbolsMap::iterator get(const std::string& name);
-	void set(const std::string& name, std::variant<float, int, bool> value);
+	void set(const std::string& name, std::variant<double, int, bool> value);
 	bool remove(const std::string& name);
 	
 	SymbolsMap symbols;
